dynamiclly render UI based on user session with SSR with next.js client component
