+++ app/auth-button-client.tsx
@@ -1,0 +1,36 @@
+"use client";
+
+import {
+    Session,
+    createClientComponentClient,
+} from "@supabase/auth-helpers-nextjs";
+import { useRouter } from "next/navigation";
+
+export default function AuthButtonClient({
+    session,
+}: {
+    session: Session | null;
+}) {
+    const supabase = createClientComponentClient();
+    const router = useRouter();
+
+    const handleSignOut = async () => {
+        await supabase.auth.signOut();
+        router.refresh();
+    };
+
+    const handleSignIn = async () => {
+        await supabase.auth.signInWithOAuth({
+            provider: "github",
+            options: {
+                redirectTo: "http://localhost:3000/auth/callback",
+            },
+        });
+    };
+
+    return session ? (
+        <button onClick={handleSignOut}>Logout</button>
+    ) : (
+        <button onClick={handleSignIn}>Login</button>
+    );
+}
