+++ app/page.tsx
@@ -1,6 +1,6 @@
 import { createServerComponentClient } from '@supabase/auth-helpers-nextjs'
 import { cookies } from 'next/headers'
-import AuthButton from './auth-button';
+import AuthButtonServer from './auth-button-server';
 
 export default async function Home() {
   const supabase = createServerComponentClient({ cookies })
@@ -9,7 +9,7 @@ export default async function Home() {
     .select();
   return (
     <>
-      <AuthButton />
+      <AuthButtonServer />
       <pre>
         {JSON.stringify(tweets, null, 2)}
       </pre>
